Updated setup.py for windows builds
When attempting to install pdftotext via pip install on Windows, there are several missing dependency errors.
This code specifically fixes the issue of missing poppler-document.h and missing poppler-cpp.lib.  
These files exist after a conda install poppler, though they are not found in the include or library directories that are being used.

This also quits if CONDA_PREFIX is None, which means that conda is not installed, or the environment variable has been whacked.